feat: allow unfree redistributable packages
We can't allow all unfree packages as we are caching to cachix, but we can
allow any redistributable ones. 

We do this by passing settings into nixpkgs. We don't want to pass settings
into non-nixpkgs inputs, so we'll need to also provide a lib option to keep
from defining an option.