feat: add helix config
We've been using helix editor since the start but up until now it has
only been in configuration.nix rather than getting config through
home-manager. The config I've added here is a compilation of a couple of
pieces of work...

- My currently-set options
- Some past config I had for some nice keybindings

Additionally, the keybinding piece also needs a long-awaited helix PR -
we can just go ahead and patch that in

---

I've kept helix in configuration.nix too as I want to make sure that we
still have a nice editor for, say, editing files as root/etc. Probably
it should go in some file in "common" instead but that can be a followup