feat: make helix display some whitespace
It's nice to be able to see whether you are using tabs, spaces, etc. in
a file and since as by default all of them display invisibly you need to
print some other character there to show you.

Displaying all of the whitespace all of the time is not a great idea...
it's distracting: both having all types of whitespace highlighted
and the default theme which displays whitespace in the same color as
comments need to be toned-down to make this usable.

We can disable some sorts of whitespace display and change the color to
be a bit dimmer to fix this issue.

We also need to either turn off indent guides or display a tabpad
character or tabs where we have indent guides will never be shown - this
wouldn't be a problem if we were showing spaces but as we aren't the
differentiator is removed.