fix(ssh-agent): add SSH_ASKPASS
Previously we enabled the ssh-agent for users but didn't add any askpass
command. This meant that if you were using a key that needed a pin it
would immediately fail.

We can fix that by providing an SSH_ASKPASS environment variable to
the service as it is started. Setting this for shells (e.g. via NixOS'
programs.ssh.enableAskPassword) doesn't work as the agent never gets
access to the set variable