feat: add swaync
swaync is a notification daemon - so finally we can stop Firefox, etc.
from creating its own windows and get notifications from things like
Thunderbird

last iteration we had problems with mako, so we were considering dunst
or swaync - swaync won out for its better (in our humble opinions)
notification center, which is closer to what we want our dashboard
concept to be, its do-not-disturb support, etc. We've added the
dashboard to Mod+N (for "Notifications")

if we ever get another window manager without a notification daemon it
may be worth pulling this out into an ingredient, but for now it's small
enough to live in niri