feat(niri): add quickshell
Quickshell is pretty neat. It allows you to write some widgets (similar
to AGS or eww), however the code to do that seems a lot more beautiful
(in my subjective opinion) than using any of those options.

To start off with, let's write a simple clock/status program - because
we are currently lacking an indicator for things such as time, battery,
etc.

I've included a few pieces:
- Temperature/weather
- Calendar events

that are not yet hooked up but which are planned for by Freshly Baked in
the future...

At the moment this is quite tightly coupled to niri - I doubt that's
going to change much because of the position we've chosen (within the
overview) which pretty much only exists in niri