feat: add portable ingredient for laptops
Some of our systems are desktop systems which don't have batteries.
Others, however, are laptops (or other portable devices) and therefore
have things to run even when not connected to power/etc.

As we're adding quickshell, we would like to display a battery indicator
for which we need upower. Let's install that only on our portable
devices