fix: freeze lix to (vulnerable) 2.93.0
CVE-2025-46415, CVE-2025-46416, CVE-2025-52991, CVE-2025-52992,
and CVE-2025-52993 were fixed in the previous lix release. That's great!

Unfortunately, not so great, is the introduction of the following bug:
- https://lix.systems/blog/2025-06-27-lix-critical-bug/

The sandbox escape bug is not particularly important to us as we run
the outputs of derivations we build - they aren't truly "untrusted" -
however the "break your system sometimes" bug is critical

Therefore, freeze lix at a known safe version until the fix for this bug
is released...