fix: eliminate duplicate fetching for eval
Previously we were fetching various things to evaluate our configuration

We were running into a bug where things would be re-fetched continuously
if they were specified by git hash, even if they hadn't changed...

Some of it was collabora-gtimelog, which I removed by adding it to npins

The remainder was niri cargo lock dependencies, which are a little
harder to fix. This said, we are currently tracking unstable niri which
we don't really need to be doing: I've read through the changes from
niri v25.05.1...main and I don't believe that any of them are important
to us at the moment

Therefore, tracking nixpkgs niri fixes this for us

---

When niri gets significant updates we may want to revisit whether to use
nixpkgs unstable on our systems/fetch only niri from unstable/etc.