feat: add nix-index
I need to find a particular path in nixpkgs - however the tool that
does this (nix-index) is for some reason not installed by default. Let's
install it along with the community database (to avoid generating things
ourselves...)

I've made a new ingredient for this - I'm not entirely set on that. It's
likely that we should move this into 'development' - or even 'packetmix'
- particularly as it contains comma and better command-not-found
handling - myself and Coded should probably discuss that, which didn't
happen here as this is effectively a drive-by patch for a thing I
quickly need...