feat(redhead): round window corners
I just updated the display on my framework laptop to the 2.8k one, and
on it the corners are rounded. It makes sense, therefore, for me to
round window corners to match. Doing this only for redhead means adding
a new ingredient just for it, which also means making another home...

I have no specifications for how much the corners should be rounded -
but the top corners seem to be about 8px with the bottom corners about
4px

I'm not entirely sure how good this'll look on external displays - I may
still change this later