feat(niri): add background to overview
The grey background is... fine. But it would be nice to have something
that more cohesively matches the rest of the system. If we blur and
darken the desktop wallpaper we'll get something that looks fine with
most configurations - though we should consider figuring out how to do
text contrast with quickshell on lighter backgrounds (or if we need to-
my background is pretty light and this looks great)