feat(teal): add stalwart
We've been using mailcow as a mailserver on midnight for a long while,
but that is dockerized.

We've also used mailu which has the same problem. Stalwart is the first
example we've seen of a mailserver that is configurable in nix and
yet has the option to configure some things (such as users) via a web
dashboard.