feat(teal): upgrade stalwart
We want to use at least stalwart 12.0, which added DAV features. This is
not yet available in 25.05. When we switched to unstable we faced a lack
of data migration - so we're now using 13.1

We still want to use the unstable module/package since as some changes
were made to the module before the current unstable version and these
are preserved to 13.1