fix(teal): correct stalwart certificate loading
Stalwart was previously given bare paths to certificates, but when using
paths to load certificates stalwart needs them in a template...

Fixing this left another issue apparent: stalwart couldn't access its
certificates as we hadn't set them to a group it could read. This wasn't
noticable in testing the web dashboard but became immediately apparent
when reading the logs for another issue.