feat(redhead): add nextcloud development server
I needed a nextcloud server for work, so time to add it to redhead...

It's unfortunately-involved to set up a nextcloud server, but this was
made much easier by taking work from the unmerged nextcloud-for-teal PR:
https://github.com/FreshlyBakedCake/PacketMix/pull/164

I don't want this on all the time - so I've hooked it up so that it
starts and stops with a systemd target. If I run any other servers on
here which need postgres, nginx, etc. then I will have to do that again