feat(copyparty): authenticate with tailscaleAuth
If we're accessing copyparty over tailscale we already know who is
accessing it. Therefore, we don't need to login again

We also need to make sure to listen on our tailscale address for
anything that is using tailscaleAuth. Otherwise the internal.files
listener will override them since as they have a less specific listen
address...