feat(copyparty): auto-allocate user volumes
We currently have users logging in via idp, but their volumes would
still need to be made manually.

Let's automatically allocate a user volume for every user that logs in

Doing so means that we should move the volumes into a different folder
which we can shadow to avoid us inheriting permissions from / (and
making everyone's folders world-readable) if our users db is lost. This
technique is from the copyparty docs:

https://github.com/9001/copyparty/blob/hovudstraum/docs/idp.md#by-default-idp-volumes-are-forgotten-on-shutdown

Even though we already save over restart, the problems stemming from
publicizing everyone's files would be critical: better not to run that
risk in the first place