feat(minion): change prompt color
On Chimera I had my prompt be pink. This was helpful for knowing if I
was on ssh to a server (and therefore shouldn't shut down) or if I was
on a local machine. Also useful for knowing if I'm in a place where I,
say can't Just Run `ssh midnight` and have it work.

I'd like to bring it back.