feat: add jj config
I've been threatening to do this for a while but it's finally time: I
have moved my jj config into development

This is mostly shared with coded's config since as I wrote the base for
both of our configs. Therefore, I asked him for his config too and
merged them - which was pretty easy. I've added emails and usernames for
everyone who uses jj in their user too.

This does change jj behavior for @celestial-starfruit but I believe the
config we have is better than default - PacketMix was always meant to be
quite opinionated ;)