feat(teal): add routes to midnight
We're going to switch over from sending traffic to midnight to sending
it to teal. We still have a couple of things on midnight that we need to
keep for a short amount of time. Therefore, we need to add routes back
to them from teal to keep them running for now...