feat(jujutsu): add permitted SSH signers list
We have SSH keys for signing some commits, and it would be nice to have
jujutsu automatically register them. This supersedes Coded's similar
list as I adapted it from the allowed signers he already had

I've nixified it so that we have the principals as a list and I've added
my own SSH signing key (just the one used for redhead as that's the one
mentioned as a signer in my config - currently I'm not sure what to do
about the lack of jj-vcs/jj#6688)

I've put it in a freshlybakedcake ingredient because it is a thing that
people may consider if they want if they're not us - I don't believe any
of the keys to be *wrong* but it is another measure of trust...

I haven't copied user keys here, if you have keys that you use to sign
commits I advise you put them here yourself

I think we should add keys from outside PacketMix if we verify that they
are correct (e.g. by getting the person who owns the email to say so
over a trusted channel or seeing the "verified" badge for a particular
key on GitHub)

Co-Authored-By: Samuel Shuert <me@thecoded.prof>