feat(jujutsu): use scriptfs for dynamic signing
As seen in jj-vcs/jj#6688, there is no builtin way to dynamically
determine which key to sign with

Unfortunately, being security key users, both myself and Coded want to
have different keys with different paths on different computers

In the previous commit, we added scriptfs, a filesystem which lets you
run a script when reading a file - and return the stdout as the contents
of that file - it's time to put that to good (and/or evil) use. We can
use it to ask the ssh-agent about our ssh keys and select whatever we
prefer (I've ordered security keys first because I want to use them
whenever available)

This is made more useful by e0ebae7ff7cce667e95c0d1c2976e88f64f291fd,
but that is currently only in my ingredient. It may be worth porting
the ssh-agent-add service it provides to a different ingredient (perhaps
remote?)