fix(stalwart): quote http.url
We previously hadn't put http.url in single quotes since as that wasn't
needed on the web UI. Unfortunately, it seems that in the configuration
this is interpreted as an expression, breaking the URL if we don't put
it in single quotes.

Coded did the investigation and I wrote the change here.

Co-Authored-By: Samuel Shuert <me@thecoded.prof>