fix(copyparty): make tailscale auth work again
This is a regression from upuvkmrlstxkxpptssslvyrvvwnwosmt (commit
4de3921b846818357f0b79960b575b7db3f71043)

In it, we changed the default listener IP to include tailscale IPs, but
this meant that the main copyparty listener was used rather than the
tailscale-authed one. Explicitly specifying that the tailscale-authed
listener should listen on the tailscale IP and that the main listener
should only listen on 0.0.0.0 fixes this.