feat(jujutsu): add jj gerrit upload
There's a jujutsu patch in the works to add a command to upload changes
to gerrit (jj-vcs/jj#7245). We use gerrit, so it's worth having this on
our copy of jujutsu.

We've used the proposed "upload" name rather than send - mostly to try
it out but also because we can't see any objections other than "the WIP
used send for so long that we all got used to it" so suspect it may get
changed...