feat(ingredients): extend to homes
Ingredients are a little tricker to use with homes than systems. That's
because in Nilla home arguments aren't special args (to allow their
portability to systems). Because this can't be changed, we previously
had special imports where we passed in inputs that needed to be
imported. As ingredients auto-import, this wouldn't be doable without
some sort of manual override functionality...

...instead of this, I've added a way to override some arguments that are
passed into ingredients. Doing this allows us to swap out our "project"
argument for a version that *can* be imported from/etc.

After doing this, moving homes to ingredients is a trivial refactor