fix(niri): start walker with a systemd service
Coded added a stopgap solution for walker starting before elephant. We
were still getting pretty bad startup performance, though, so I looked
into it.

It turned out a variety of things were needed to start walker faster:
- Moving to a systemd service was helpful, but only once walker had
  started once, otherwise the socket would never connect and walker
  wouldn't open at all
- Disabling file indexing by default was necessary for systems with lots
  of files in home (for example LibreOffice clones). If you want to add
  file indexing back for yourself, I recommend you figure out where your
  large files are and use '.ignore' files (.gitignore syntax) to ignore
  them
- Elephant needed to be started properly - after niri so WAYLAND_DISPLAY
  was set but not in the ordering cycle on graphical-session.target. I
  used the same settings as we have for ssh-agent, say

Because of all of this, I also factored walker out into its own file.

This additionally marks

- The first keybind moved outside of the niri.nix file
- The first keybind to be created with Mod rather than ${mod} (= Super)

While it's a bit of a shame to not have all the keybindings in the same
place, I think it's worth it...