fix(spindle): open firewall for all interfaces
There's a race condition here where teal sometimes looks up midnight
when not connected to tailscale. If it does that, it resolves midnight
on the local network. That would be entirely fine if we weren't just
listening on Tailscale.

Further, that lookup can then get cached, bringing down the spindle even
when tailscale comes up

We trust the local network, let's just allow this route too...