fix(pm/minion/p10k): fix async worker restarting
Previously we were using a subshell to group our commands, which can't
be used to run an async worker as the async worker is tied to a shell.
Instead, we should be using a command group. We also need to silence
errors on the first job invocation or it'll still print out