feat(pm/redhead): add nextcloud-docker-dev
I'm using https://github.com/juliusknorr/nextcloud-docker-dev to test
some changes to a nextcloud plugin, as the plugin needs a more recent
version of Nextcloud than is in nixpkgs.

To be available, I need to set up my hosts file correctly and forward
this over nginx for my other devices.